Define atom for ART counters
This CL defines the atom used by ART to encode fine grained metrics
about ART's behavior and performance.

ART keeps a set of internal counters. A subset of these are exported to
statsd. The list of items exported by ART is given in the ArtDatumId
enum, and the values of these items will be encoded in an
ArtDatumReported message. These reports are grouped into sessions, which
indicate that the metrics all were reported by the same ART runtime
instance.

Bug: 170149255
Test: m
Change-Id: I6f987cbb1694abc8c1658218362622bd09ebac08