New metric: Log requiring of password complexity.
SET_PASSWORD_COMPLEXITY
A profile owner or a device owner may set password requirements in the
form of a platform-defined password complexity level (none, low, medium
or high). See DevicePolicyManager.setRequiredPasswordComplexity for more
details.
We log this metric when the password complexity requirement was set to a
value different to a previous password complexity (or quality).
We log the administrator app package, the complexity level required, and
whether it was applied device-wide or only for the managed profile.

Bug: 165573442
Test: atest com.android.cts.devicepolicy.MixedDeviceOwnerTest#testPasswordMethodsLogged
Change-Id: Iae040027a4c40c52eab5970089e35c8145396b35