Add cred mng app metrics enums
PDD: https://eldar.corp.google.com/assessments/875507949/drafts/232953480?jsmode=o

CREDENTIAL_MANAGEMENT_APP_REQUEST_NAME
We log the package name of the app requesting to
manage the user's KeyChain credentials on
an unmanaged device.
The log data consists of the requesting app's package name.

CREDENTIAL_MANAGEMENT_APP_REQUEST_POLICY
We log the number or apps and URIs in the authentication
policy of the app requesting to manage the user's
KeyChain credentials on an unmanaged device.
The log data consists of the number of apps and number of URIs
in the requesting app's authentication policy.

CREDENTIAL_MANAGEMENT_APP_REQUEST_ACCEPTED
We log this when the user accepts a request to allow
an app to manage their KeyChain credentials.
There is no log data (only the event).

CREDENTIAL_MANAGEMENT_APP_REQUEST_DENIED
We log this when the user denies a request to allow
an app to manage their KeyChain credentials.
There is no log data (only the event).

CREDENTIAL_MANAGEMENT_APP_REQUEST_FAILED
We log this when a request made to manage the user's
KeyChain credentials fails. A request will fail
if the device is a managed device or if the
requesting app's package name and policy is invalid.
There is no log data (only the event).

CREDENTIAL_MANAGEMENT_APP_CREDENTIAL_FOUND_IN_POLICY
We log this when a KeyChain credential is successfully
found in the credential management app's
authentication policy. This credential is presented
to a remote service for authentication.
There is no log data (only the event).

CREDENTIAL_MANAGEMENT_APP_INSTALL_KEY_PAIR_FAILED
We log this when the credential management app fails
to install a key pair on the device.
The log data consists of the credential management app's
package name.

CREDENTIAL_MANAGEMENT_APP_GENERATE_KEY_PAIR_FAILED
We log this when the credential management app fails
to generate a key pair on the device.
The log data consists of the credential management app's
package name.

CREDENTIAL_MANAGEMENT_APP_POLICY_LOOKUP_FAILED
We log this when the check for whether a KeyChain
credential is found in the credential management app's
authentication policy fails.
There is no log data (only the event).

CREDENTIAL_MANAGEMENT_APP_REMOVED
We log this when the user removes the credential
management app from Settings.
There is no log data (only the event).

Bug: 165641221
Test: none
Change-Id: I4ecc0a20a14a3c9ed3483b312d66c1998c101f2f