New metric: Setting Organization ID
SET_ORGANIZATION_ID
A profile owner or a device owner may set the organization identifier,
identifying the organization the device or work profile was enrolled
into, and triggering the generation of an enrollment-specific ID.

See DevicePolicyManager.setOrganizationId for more details.

We log this metric when the organization ID is set, as a proxy for
adoption of the Enrollment-specific ID.
value different to a previous password complexity (or quality).
We log the administrator app package and whether the ID was set on a
fully-managed device or on a work profile.
Notably, we do *not* log the Organization ID itself.

Bug: 168627890
Test: atest com.android.cts.devicepolicy.MixedDeviceOwnerTest#testEnrollmentSpecificIdCorrectCalculation com.android.cts.devicepolicy.MixedManagedProfileOwnerTest#testEnrollmentSpecificIdCorrectCalculation
Change-Id: I7a3d85f0f0c4ed4051f25d0b45b3fdccd2884bea