Make stats-log-api-gen available as a prebuilt
This involved two steps. First, we make stats-log-api-gen statically
link all of its dependencies because the SDK prebuilt snapshot
generation currently does not support host tools with dynamic
library dependencies (b/178691801)

Second, we add a module_exports entry to Android.bp, which will
actually make stats-log-api-gen available as a prebuilt to modules
that need it. This is done in a separate CL:
https://r.android.com/1573375

Bug: 178500083
Bug: 178691801
Test: Copy stats-log-api-gen out of tree and run it there.
Change-Id: I2dfa215368beeeab73de7732596ff38715d64542