Fix the the proto definition for app startup and fully drawn events.
ActivityMetricsLoggers build the proto with the compilation reason first
and then the compilation filter. Instead of changing the logging code,
we prefer to change the proto field order: this will facilitate the
speed up the resolution without the need to purge data or wait for
further release builds.

It is ok, because both fields are in fact int32 proto values so the ABI
does not change.

Test: m
Bug: 178108751
Change-Id: I03c701efe2181ae7796547267902f884007593eb