Add Metrics enum item for Network Slicing change
This metrics is to collect the status of network slicing status through SET_NETWORK_SLICING_ENABLED, which is a boolean flag to indicate whether 5g slicing is enabled on the work profile. Slicing allows operators to virtually divide their networks in portions and use different portions for specific use cases; for example, a corporation can have a deal/agreement with a carrier that all of its employees’ devices use data on a slice dedicated for enterprise use.

Test: Treehugger
Bug: 178655595
Change-Id: I7293c97381ae30fc0ac4ff87b5ecd9c8a6717255
Merged-In: I7293c97381ae30fc0ac4ff87b5ecd9c8a6717255