MediaMetrics: Add log session id to audio components
In order to synchronize low level media components to Android
Media Playback-level Metrics, we are adding a log_session_id
field into the AudioTrack and AudioRecord protos.

The log_session_id is a 128 bit randomly generated integer
encoded as a Base64Url string, which can be utilized to
connect associated playback components together.

Test: compiles with matching change in frameworks/av
Bug: 181032765
Change-Id: Id89175483776a76060c2e2f1297b8500ff3ecb17