Move network transport to proto_logging
Move the network transport type to proto_logging. It is needed in
atoms.proto, and all atoms.proto dependencies should live in
proto_logging. Previously, networkcapabilities.proto was in
proto_logging, but the majority of that proto is not needed for logging
and is used for incident reports.

Test: m
Bug: 182392879
Change-Id: I4da9b12d87fb3633c7b2f959175702bfc0a92073
Merged-In: I4da9b12d87fb3633c7b2f959175702bfc0a92073