Add metrics for landscape and night mode
PROVISIONING_IS_LANDSCAPE
Logs whether a screen is in landscape mode. This metric is
logged for each individual ManagedProvisioning screen. It is
also logged upon screen rotation, in order to monitor usage of
the new landscape support. The log data consists of a boolean
whether the screen is in landscape mode, and the class name
of the ManagedProvisioning screen.

PROVISIONING_IS_NIGHT_MODE
Logs whether ManagedProvisioning is in dark mode. This metric
is only logged once for the lifetime of ManagedProvisioning.
This metric is necessary in order to monitor usage of the
new night mode support. The log data only consists of a
boolean which is true if the app is using night mode.

Fixes: 183032260
Test: none
Change-Id: I354b2ed5ed4e6ac04ee84ed86d3cfcc61f0a10a5