BatteryUsageStats atom
Adds a new atom for tracking battery usage, per uid as well
as for the device, broken down by components. This loosely
follows the structure of the BatteryUsageStats class.

There are in fact three new atoms, all with the same structure,
representing different time periods and calculation methods
in the BatteryStats mechanism.

Bug: 184095105
Bug: 187996428
Test: atest BatteryUsageStatsProtoTests
Change-Id: I75a900ea185c5db18dc6f07d31f6d73eb7bbbb48