Add new atom ANROccurredProcessingStarted
The atom will be logged immediately after an ANR occurs. This atom will
only be logged with the purpose of triggering Perfetto and the logging
needs to happen as close as possible to the time when the ANR is
detected. Unfortunately, that is not possible with the current atom
(ANROccurred) since another issue requires ANROccurred to be logged
after the ANR dump to the filesystem (see ag/14392119).

Metric Design Review: eldar/845311492
Bug: b/188122403
Test: Manual (`statsd_testdrive 373`).

Change-Id: I3c766a966a7643866b3b1c3db04394f8014319fc