Move start_time_millis to the outermost atom.
The start_time_millis field is required to slice the I/O usage summary
metrics by metrics event start date. This field can be used to slice
the metrics only when it is logged directly and not as bytes. Ergo, Move
the field out of the message that is logged as bytes.

Test: Build is okay. Tested with statsd_testdrive.
Bug: 216524394

Change-Id: I6f2f2ad4e02a9b6752764b8ccd215291f213a810
Merged-In: I6f2f2ad4e02a9b6752764b8ccd215291f213a810